STRAT-17 - State of Charge Investigation Preliminary
SoC Algorithm based on the energy left in the pack. Not sure how we want to remove energy from the virtual pack, so made a basic function to do so.

Also a basic model for cell efficiency and associated ohmic power losses

Need to talk with Clarke about how this will fit with everything.

Note - not tested or anything, just programmed it in Notepad++