[STRAT-82] Create backend API that will return mock data (#54)
* [STRAT-53] Show energy level over the course of a race (#45)

* [STRAT-53][Show energy level over the course of a race] Creates a graph of state of charge and velocity against distance travelled. Add functions to generate desired SOC values given velocity and elevation profiles. Uses Car and ColoumbCounter objects to calculate energy used and state of charge.

* Refactored, tests, and README updated (#46)

* Add some information to Getting Started section to help explain some steps for people new to development. (#48)

* Basic Flask API endpoint.

* Add processing.py which performs GET request to some dummy API endpoint,
parses it into the desired format, and tries calling some other model we
have to perform some operation on it.

* Add constant to store API endpoint.

* Redundant flask imports in api.py.

* Configuring existing API funcitonality to include database.

* Trying to find out order of imports to fix database operational error.

* Functional - produces the expected output into a db.sqlite file. Fix was
that the basedir of the database was wrong, I was using sys.path[0]
instead of os.path.abspath(os.path.dirname()). Also to be fixed is the
file structure, right now everything is in the app.py file which while
functional is not ideal.

* Restructure api folder so that global variables available through
__init__, renamed log_models to models, and app.py now queries through
db to fetch data to make request. In current form, still done upon a
request made, but want to separate that in a later commit.

* Add timer using APScheduler BackgroundScheduler. Deleted processing.py
as function used was moved to app.py.

* Update README.md with endpoints.

Co-authored-by: ufshk <ufsheikh@live.ca>